cleanup, move screenshot into hook
